+++ ext/duckdb/appender.c
@@ -9,10 +9,10 @@ static VALUE appender_initialize(VALUE klass, VALUE con, VALUE schema, VALUE tab
 static VALUE appender_error_message(VALUE self);
 static VALUE appender__append_bool(VALUE self, VALUE val);
 static VALUE appender__append_int8(VALUE self, VALUE val);
-static VALUE appender__apend_int16(VALUE self, VALUE val);
-static VALUE appender_append_int32(VALUE self, VALUE val);
-static VALUE appender_append_int64(VALUE self, VALUE val);
-static VALUE appender_append_uint8(VALUE self, VALUE val);
+static VALUE appender__append_int16(VALUE self, VALUE val);
+static VALUE appender__append_int32(VALUE self, VALUE val);
+static VALUE appender__append_int64(VALUE self, VALUE val);
+static VALUE appender__append_uint8(VALUE self, VALUE val);
 static VALUE appender_append_uint16(VALUE self, VALUE val);
 static VALUE appender_append_uint32(VALUE self, VALUE val);
 static VALUE appender_append_uint64(VALUE self, VALUE val);
@@ -138,7 +138,7 @@ static VALUE appender__append_int8(VALUE self, VALUE val) {
 }
 
 /* :nodoc: */
-static VALUE appender__apend_int16(VALUE self, VALUE val) {
+static VALUE appender__append_int16(VALUE self, VALUE val) {
     rubyDuckDBAppender *ctx;
     int16_t i16val = (int16_t)NUM2INT(val);
 
@@ -147,72 +147,34 @@ static VALUE appender__apend_int16(VALUE self, VALUE val) {
     return duckdb_state_to_bool_value(duckdb_append_int16(ctx->appender, i16val));
 }
 
-/* call-seq:
- *   appender.append_int32(val) -> self
- *
- * Appends an int32(INTEGER) value to the current row in the appender.
- *
- *   require 'duckdb'
- *   db = DuckDB::Database.open
- *   con = db.connect
- *   con.query('CREATE TABLE users (id INTEGER, age INTEGER)')
- *   appender = con.appender('users')
- *   appender
- *     .append_int32(1)
- *     .append_int32(20)
- *     .end_row
- *     .flush
- */
-static VALUE appender_append_int32(VALUE self, VALUE val) {
+/* :nodoc: */
+static VALUE appender__append_int32(VALUE self, VALUE val) {
     rubyDuckDBAppender *ctx;
     int32_t i32val = (int32_t)NUM2INT(val);
 
     TypedData_Get_Struct(self, rubyDuckDBAppender, &appender_data_type, ctx);
 
-    if (duckdb_append_int32(ctx->appender, i32val) == DuckDBError) {
-        rb_raise(eDuckDBError, "failed to append int32");
-    }
-    return self;
+    return duckdb_state_to_bool_value(duckdb_append_int32(ctx->appender, i32val));
 }
 
-/* call-seq:
- *   appender.append_int64(val) -> self
- *
- * Appends an int64(BIGINT) value to the current row in the appender.
- *
- *   require 'duckdb'
- *   db = DuckDB::Database.open
- *   con = db.connect
- *   con.query('CREATE TABLE users (id INTEGER, age BIGINT)')
- *   appender = con.appender('users')
- *   appender
- *     .append_int32(1)
- *     .append_int64(20)
- *     .end_row
- *     .flush
- */
-static VALUE appender_append_int64(VALUE self, VALUE val) {
+/* :nodoc: */
+static VALUE appender__append_int64(VALUE self, VALUE val) {
     rubyDuckDBAppender *ctx;
     int64_t i64val = (int64_t)NUM2LL(val);
 
     TypedData_Get_Struct(self, rubyDuckDBAppender, &appender_data_type, ctx);
 
-    if (duckdb_append_int64(ctx->appender, i64val) == DuckDBError) {
-        rb_raise(eDuckDBError, "failed to append int64");
-    }
-    return self;
+    return duckdb_state_to_bool_value(duckdb_append_int64(ctx->appender, i64val));
 }
 
-static VALUE appender_append_uint8(VALUE self, VALUE val) {
+/* :nodoc: */
+static VALUE appender__append_uint8(VALUE self, VALUE val) {
     rubyDuckDBAppender *ctx;
-    int8_t ui8val = (uint8_t)NUM2UINT(val);
+    uint8_t ui8val = (uint8_t)NUM2UINT(val);
 
     TypedData_Get_Struct(self, rubyDuckDBAppender, &appender_data_type, ctx);
 
-    if (duckdb_append_uint8(ctx->appender, ui8val) == DuckDBError) {
-        rb_raise(eDuckDBError, "failed to append uint8");
-    }
-    return self;
+    return duckdb_state_to_bool_value(duckdb_append_uint8(ctx->appender, ui8val));
 }
 
 static VALUE appender_append_uint16(VALUE self, VALUE val) {
@@ -458,9 +420,6 @@ void rbduckdb_init_duckdb_appender(void) {
     rb_define_alloc_func(cDuckDBAppender, allocate);
     rb_define_method(cDuckDBAppender, "initialize", appender_initialize, 3);
     rb_define_method(cDuckDBAppender, "error_message", appender_error_message, 0);
-    rb_define_method(cDuckDBAppender, "append_int32", appender_append_int32, 1);
-    rb_define_method(cDuckDBAppender, "append_int64", appender_append_int64, 1);
-    rb_define_method(cDuckDBAppender, "append_uint8", appender_append_uint8, 1);
     rb_define_method(cDuckDBAppender, "append_uint16", appender_append_uint16, 1);
     rb_define_method(cDuckDBAppender, "append_uint32", appender_append_uint32, 1);
     rb_define_method(cDuckDBAppender, "append_uint64", appender_append_uint64, 1);
@@ -480,7 +439,10 @@ void rbduckdb_init_duckdb_appender(void) {
     rb_define_private_method(cDuckDBAppender, "_close", appender__close, 0);
     rb_define_private_method(cDuckDBAppender, "_append_bool", appender__append_bool, 1);
     rb_define_private_method(cDuckDBAppender, "_append_int8", appender__append_int8, 1);
-    rb_define_private_method(cDuckDBAppender, "_append_int16", appender__apend_int16, 1);
+    rb_define_private_method(cDuckDBAppender, "_append_int16", appender__append_int16, 1);
+    rb_define_private_method(cDuckDBAppender, "_append_int32", appender__append_int32, 1);
+    rb_define_private_method(cDuckDBAppender, "_append_int64", appender__append_int64, 1);
+    rb_define_private_method(cDuckDBAppender, "_append_uint8", appender__append_uint8, 1);
     rb_define_private_method(cDuckDBAppender, "_append_date", appender__append_date, 3);
     rb_define_private_method(cDuckDBAppender, "_append_interval", appender__append_interval, 3);
     rb_define_private_method(cDuckDBAppender, "_append_time", appender__append_time, 4);