+++ ext/duckdb/logical_type.c
@@ -5,8 +5,10 @@ static VALUE cDuckDBLogicalType;
 static void deallocate(void *ctx);
 static VALUE allocate(VALUE klass);
 static size_t memsize(const void *p);
+static VALUE duckdb_logical_type__type(VALUE self);
 static VALUE duckdb_logical_type_width(VALUE self);
 static VALUE duckdb_logical_type_scale(VALUE self);
+static VALUE duckdb_logical_type_child_type(VALUE self);
 
 static const rb_data_type_t logical_type_data_type = {
     "DuckDB/LogicalType",
@@ -35,6 +37,19 @@ static size_t memsize(const void *p) {
 
 /*
  *  call-seq:
+ *    decimal_col.logical_type.type -> Symbol
+ *
+ *  Returns the logical type's type symbol.
+ *
+ */
+static VALUE duckdb_logical_type__type(VALUE self) {
+    rubyDuckDBLogicalType *ctx;
+    TypedData_Get_Struct(self, rubyDuckDBLogicalType, &logical_type_data_type, ctx);
+    return INT2FIX(duckdb_get_type_id(ctx->logical_type));
+}
+
+/*
+ *  call-seq:
  *    decimal_col.logical_type.width -> Integer
  *
  *  Returns the width of the decimal column.
@@ -59,6 +74,34 @@ static VALUE duckdb_logical_type_scale(VALUE self) {
     return INT2FIX(duckdb_decimal_scale(ctx->logical_type));
 }
 
+/*
+ *  call-seq:
+ *    list_col.logical_type.child_type -> DuckDB::LogicalType
+ *
+ *  Returns the child logical type for list and map types, otherwise nil.
+ *
+ */
+static VALUE duckdb_logical_type_child_type(VALUE self) {
+    rubyDuckDBLogicalType *ctx;
+    duckdb_type type_id;
+    duckdb_logical_type child_logical_type;
+    VALUE logical_type = Qnil;
+
+    TypedData_Get_Struct(self, rubyDuckDBLogicalType, &logical_type_data_type, ctx);
+    type_id = duckdb_get_type_id(ctx->logical_type);
+
+    switch(type_id) {
+        case DUCKDB_TYPE_LIST:
+        case DUCKDB_TYPE_MAP:
+            child_logical_type = duckdb_list_type_child_type(ctx->logical_type);
+            logical_type = rbduckdb_create_logical_type(child_logical_type);
+            break;
+        default:
+            logical_type = Qnil;
+    }
+    return logical_type;
+}
+
 VALUE rbduckdb_create_logical_type(duckdb_logical_type logical_type) {
     VALUE obj;
     rubyDuckDBLogicalType *ctx;
@@ -78,6 +121,8 @@ void rbduckdb_init_duckdb_logical_type(void) {
     cDuckDBLogicalType = rb_define_class_under(mDuckDB, "LogicalType", rb_cObject);
     rb_define_alloc_func(cDuckDBLogicalType, allocate);
 
+    rb_define_private_method(cDuckDBLogicalType, "_type", duckdb_logical_type__type, 0);
     rb_define_method(cDuckDBLogicalType, "width", duckdb_logical_type_width, 0);
     rb_define_method(cDuckDBLogicalType, "scale", duckdb_logical_type_scale, 0);
+    rb_define_method(cDuckDBLogicalType, "child_type", duckdb_logical_type_child_type, 0);
 }