+++ lib/duckdb/result.rb
@@ -27,24 +27,10 @@ module DuckDB
     RETURN_TYPES = %i[invalid changed_rows nothing query_result].freeze
 
     alias column_size column_count
-    alias row_size row_count
 
     class << self
       def new
         raise DuckDB::Error, 'DuckDB::Result cannot be instantiated directly.'
-      end
-
-      def use_chunk_each=(value) # :nodoc:
-        raise('`changing DuckDB::Result.use_chunk_each to false` was deprecated.') unless value
-
-        warn('`DuckDB::Result.use_chunk_each=` will be deprecated.')
-
-        true
-      end
-
-      def use_chunk_each? # :nodoc:
-        warn('`DuckDB::Result.use_chunk_each?` will be deprecated.')
-        true
       end
     end